Creare modele.

+++ src/main/java/com/hospital/hospital_system/repository/AsistentRepository.java
@@ -4,7 +4,7 @@ import com.hospital.hospital_system.models.Asistent;
 import org.springframework.data.jpa.repository.JpaRepository;
 import java.util.Optional;
 
-public interface AsistentRepository extends JpaRepository<Asistent, Long> {
+public interface AsistentRepository extends JpaRepository<Asistent, Integer> {
 
     Optional<Asistent> findByCnp(String cnp);
 }
